+++ pennylane/exceptions.py
@@ -34,3 +34,7 @@ class PennyLaneDeprecationWarning(UserWarning):  # pragma: no cover
 
 class ExperimentalWarning(UserWarning):  # pragma: no cover
     """Warning raised to indicate experimental/non-stable feature or support."""
+
+
+class TransformError(Exception):  # pragma: no cover
+    """Raised when there is an error with the transform logic."""
