+++ pennylane/ops/functions/matrix.py
@@ -21,10 +21,10 @@ from typing import Union
 
 import pennylane as qml
 from pennylane import transform
+from pennylane.exceptions import TransformError
 from pennylane.operation import MatrixUndefinedError, Operator
 from pennylane.pauli import PauliSentence, PauliWord
 from pennylane.tape import QuantumScript, QuantumScriptBatch
-from pennylane.transforms import TransformError
 from pennylane.typing import PostprocessingFn, TensorLike
